crypto/zkpay/solver/src/requirements.txt
requests == 2.22.0
pillow == 6.1.0
qrcode == 6.1
zbarlight == 2.3
